feat: extract logic to CatalogClient and wrap in Suspense with spinner fallback
